Address PR review feedback for drinking water package extraction
- Rename package from movici-sim-drinking-water to movici-drinking-water-model
- Flatten test directory: move tests directly under tests/
- Move JSON schema to package root, remove json_schemas subpackage
- Remove ruff config from package pyproject.toml, consolidate in root
- Add README.rst for PyPI publishing
- Add per-package Makefile mirroring simulation-core
- Move test/coverage targets to root Makefile using package globs
- Update CI workflows to use root Makefile targets
