+++ Makefile
@@ -7,6 +7,19 @@ toml-check:
 
 lint: ruff toml-check
 
+unittest:
+	NUMBA_DISABLE_JIT=1 uv run pytest -v packages/*/tests/
+
+coverage:
+	NUMBA_DISABLE_JIT=1 uv run pytest \
+		$(patsubst %,--cov %,$(wildcard packages/*/src/*)) \
+		--cov-report=term-missing \
+		--cov-report=xml \
+		--cov-report=html \
+		packages/*/tests/
+
+test-numba:
+	uv run pytest -v packages/*/tests/
 
 .PHONY: docs
 docs:
