+++ packages/drinking-water/pyproject.toml
@@ -3,12 +3,30 @@ requires = ["hatchling", "uv-dynamic-versioning"]
 build-backend = "hatchling.build"
 
 [project]
-name = "movici-sim-drinking-water"
+name = "movici-drinking-water-model"
 dynamic = ["version"]
 description = "Drinking water network simulation model for Movici using WNTR"
+readme = "README.rst"
 requires-python = ">=3.10,<3.14"
 license = { text = "Movici Public License" }
 authors = [{ name = "NGinfra Movici", email = "movici@nginfra.nl" }]
+maintainers = [{ name = "NGinfra - Movici", email = "movici@nginfra.nl" }]
+classifiers = [
+  "Development Status :: 4 - Beta",
+  "Intended Audience :: Science/Research",
+  "Intended Audience :: Developers",
+  "License :: Free for non-commercial use",
+  "License :: Other/Proprietary License",
+  "Operating System :: Microsoft :: Windows",
+  "Operating System :: POSIX :: Linux",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Scientific/Engineering",
+  "Topic :: Scientific/Engineering :: GIS",
+]
 
 dependencies = [
   "movici-simulation-core",
@@ -18,20 +36,11 @@ dependencies = [
 ]
 
 [project.entry-points."movici.plugins"]
-drinking_water = "movici_sim_drinking_water.model:Model"
-drinking_water_attributes = "movici_sim_drinking_water.attributes:DrinkingWaterNetworkAttributes"
+drinking_water = "movici_drinking_water_model.model:Model"
+drinking_water_attributes = "movici_drinking_water_model.attributes:DrinkingWaterNetworkAttributes"
 
 [tool.hatch.version]
 source = "uv-dynamic-versioning"
 
 [tool.uv-dynamic-versioning]
 fallback-version = "0.0.0"
-
-[tool.ruff]
-extend = "../../pyproject.toml"
-
-[tool.ruff.lint.per-file-ignores]
-"tests/*" = ["S"]
-
-[tool.ruff.lint.isort]
-known-first-party = ["movici_simulation_core", "movici_sim_drinking_water"]
