+++ packages/drinking-water/src/movici_drinking_water_model/attributes.py
