+++ packages/drinking-water/src/movici_drinking_water_model/drinking_water.json
