+++ packages/drinking-water/src/movici_drinking_water_model/network_wrapper.py
