--- packages/drinking-water/src/movici_sim_drinking_water/json_schemas/__init__.py
@@ -1,5 +1,0 @@
-import pathlib
-
-SCHEMA_PATH = pathlib.Path(__file__).parent
-
-__all__ = ["SCHEMA_PATH"]
